python.lang: recognize escape sequences in raw strings
Raw strings can still contain escape sequences. In particular, the quote
character can be escaped. However, the escape sequences are interpreted
literally (i.e., as a backslash followed by other characters).
Illustrate that by suppressing the style of the escaped-char context.

Line continuations are interpreted literally inside raw strings too (as
a backslash and a newline), so suppress the style for those, as well.

https://bugzilla.gnome.org/show_bug.cgi?id=784673