docs: porting guide: remove recommendation to build 3.99.x tarballs
It was initially planned for GtkSourceView 4 to rename the namespace
from GtkSource to Gsv. But this won't happen: error, not enough funding.

With the namespace change it would have made sense to build intermediate
tarballs, to handle the small GtkSource API breaks first, and then
handling the GtkSource -> Gsv namespace change with a separate commit.
But handling all the GtkSource API breaks at once is entirely feasible,
since it's mostly renaming functions version 2 back to the original
names.

In the porting guide, keep sub-sections for each 3.99.x version, it
doesn't hurt to keep that information.