language-specs: sync all *.lang files from GtkSourceView 4
gedit is unmaintained, and still uses deprecated GtkSourceView 3 APIs.
So it's unlikely that gedit will use GtkSourceView 4 in the near future.

So from time to time the *.lang files can be synced from GtkSourceView 4
to GtkSourceView 3, so that gedit (and other GSV 3 apps) benefit from
the latest improvements for syntax highlighting, and support for new
languages.

Since it is the first time that the *.lang files are synced, a lot of
strings are no longer translatable, see the commits in GtkSourceView 4
for more details.

Normally this commit should not add new translatable strings.