groovy.lang: fix highlighting for dollar slashy strings starting with /
The look-ahead assertion is wrong; what _actually_ can't follow `$/` is
not `/`, but `/$`. In other words, a dollar slashy string can't be empty.

Also, remove the comment about the requirements on the on the preceding
token, because those were removed in the in-development versions of Groovy
(not sure when exactly, but they're gone in 3.0.0-alpha-1).

https://bugzilla.gnome.org/show_bug.cgi?id=792762