css.lang: Support Unicode and character escapes for IDs and classes
This adds an identifier regex that closely follows the rules for CSS
identifiers, and uses this regex wherever CSS identifiers are used
(ID/class/tag selectors, function names), so that Unicode characters and
character escapes are correctly highlighted.

This also removes the unicode-character-reference context, as Unicode
character escapes do not appear anywhere else (except for inside
strings, but then they are highlighted as part of the string).

https://bugzilla.gnome.org/show_bug.cgi?id=796130