completion: fix improper memory use for background-color
gtk_style_context_get() returns a boxed copy of the GdkRGBA for the given
state. Instead of providing a GdkRGBA* as the parameter, we need to provide
a GdkRGBA** and free it upon exit from the function.

This fixes a leak detected by LSAN. While the previous usage could have
resulted in incorrect visuals, it was larger than a pointer size and
therefore should not have caused any memory corruption.