groovy.lang: correctly highlight slashes after different types of strings
Slashes are interpreted as division operators (and not beginnings of
slashy strings) after string literals. I had previously implemented this
for slashes following double-quoted strings, but I evidently forgot
about other types of string literals.

Implement it for single-quoted and dollar slashy strings, as well.

There are also (non-dollar) slashy strings, but there's no way for us to
tell if the preceding token is a slashy string or a division operator,
so that case will have to remain unimplemented.