gtksourcestylescheme: re-apply CSS when parent style is set
We generate the CSS before the parent_id has been resolved into a pointer
for another style within the style manager. That means we have cached
values that are incorrect.

This updates the CSS when the parent pointer is set so that the background
and other text styles are updated correctly in the CSS provider.

Fixes #37