buffer: gtk_source_buffer_create_source_mark is transfer none
This makes our code adhere to the contract defined in the documentation
that is this function should return a borrowed reference, not a full
reference.

This function has been miss-documented since 2003 in commit
fe2df9df6741d3d484ea73243d43bad09f0d6c5b so I think it is safe to implement
it as documented.

Related to #39