testsuite: Fix a comparison which depended on GLib hash table ordering
In GLib 2.59, the iteration order of GHashTable was changed (its order
has always be, and will always be, undefined). This has broken some
comparisons in check_language() in test-language.c, which assumed a
certain hash table ordering for the results returned by
gtk_source_language_get_*().

Fix that by using an unordered comparison function. This should be
backwards compatible with older GLib versions.

Signed-off-by: Philip Withnall <withnall@endlessm.com>