utils: add fast string conversion from Builder
This provides a string data section that contains strings from 10,000 to
20,000 which can be used to convert integers from 0 to 20,000 with only
a few lookups.

If the value is > 20,000, then a temporary buffer will be used along with
a g_snprintf().

This can be used to significantly reduce the overhead of line number
calculation in tight loops such as when drawing the gutter on every
frame. Builder uses this in the OmniGutter implementation and this
consitutes a port of that into GtkSourceView.