contextengine: disable highlighting in presence of long lines
If we come across a pathologically long line (defined as 2000 characters),
then disable highlighting. As highlighting is line-based, this allows us
to protect ourselves and keep the text editor working as fast as
GtkTextView is willing to support.

In particular, this also helps mitigate against libpcre's insistance on
stack-overflow'ing in some situations from long input combined with
complex regexes.

It would be nice if we could determine what type of regexes are likely to
be a potential failure for this and only limit those, but this solution
is more likely to have the secondary effect of improving performance for
minified HTML, CSS, etc.

I think a pre-requisite to remove this feature would be for GLib to move
to using libpcre2 for GRegex (also allowing for JIT support) as well as
GtkTextView improving beyond line-based layout/rendering (which probably
requires character indexes within lines).

Fixes #47