build: bump GTK requirement to >= 3.22
This bumps the GTK requirement to 3.22 so that we can start the process
of removing deprecations. For distributions wanting updates to
GtkSourceView without the ability to use GTK 3.22, the gtksourceview-4-2
will continue to provide that for you.

However, primary development must now start to think about what a
GtkSourceView will look like with GTK 4.x, and the first step is to
start removing deprecations.