yaml.lang: improve detection of block scalars
Allow chomping indicators, indentation indicators, trailing whitespace
and a trailing comment on the line with the block scalar indicator.
See the YAML 1.2 spec, section 8.1.1.

Unfortunately, we can't use the indentation indicator to determine which of
the following lines to include in the scalar, so a block scalar with an
indentation indicator will still be highlighted incorrectly overall, unless the
indicator is redundant.