perl.lang: Fix regular expression highlighting in statement modifiers
This fixes regular expression highlighting for all statement modifiers
that accept an expression (if/unless/while/until/when), and highlights
the modifier as a keyword instead of builtin.

This also updates the lists of modifiers for regular expressions. (The
lists are from perlop 5.30.0[1], and are copied directly with no
alphabetical sorting.)

Fixes #74.

[1]: https://perldoc.perl.org/5.30.0/perlop.html