classic: improve grid color styling
The classic style is used by default and therefore needs to have
a reasonably good look on a number of GTK themes. This uses an
alpha channel to draw the grid so that we are less likely to
have a garish view.

The style is .1 alpha of the draw-spaces color.

Fixes #98