undo: remove GtkSourceUndoManager
Undo is now managed by GTK itself (with a more memory compact form of
the undo manager from GtkSourceView). It has the added benefit that it
is run from the default GtkTextBuffer handlers right next to the code
that mutates the B-Tree, meaning that we have a better chance at avoiding
reply/reordering issues which can happen today.