completion: remove GtkSourceCompletionContainer
This is unfortunate, but we cannot subclass GtkScrolledWindow anymore. We
might bring this back, but more likely is that we'll reintroduce the
sizing feature as part of a modernized completion API using newer GLib
features such as GListModel.

The fallout from this is tight control over the completion window height
is temporarily lost until the new API lands.