tango.xml, classic.xml: tweak line-number styling
Since Adwaita 3.24, the styling for the gutters has been less dramatic.
As this GtkSourceView style scheme does not specify a line-number
background, very little contrast was available.

Since this style-scheme needs to be rather agnostic in terms of GTK
theme, we can use an RGBA style to overlay similar to the right margin.
It's not perfect, but it fixes the typical case without causing too
much pain for the other cases.

In the future we may want to provide a strategy for GTK theme agnostic
style-schemes to pull values from the GTK theme directly (or mix() between
values).

Fixes #102