fix const'ness of search-path usage in various managers
These should be "const gchar * const *" since we're breakin API/ABI
anyway. This is not an ABI break, but it will cause warnings in existing
code that was passing "gchar**". This is more correct however since we
do not take any references or mutate the strings, or pointer to strings.