fonts: bundle BuilderBlocks
If we have fontconfig and pangoft2, we can override the font that is used
within the GtkSourceMap to be our "block" font. This is much more readable
than the scaled down 1pt font we fallback to. It is also faster to raster.

This, in practice, will only work where those libraries are available but
certainly covers Linux and BSD. Possibly covers Mac (needs testing) and
optionally covers Windows.

Other font loaders would be happily accepted.