map: work around recent gtk theming API changes
We lost a bunch of theming API in GTK 4, so we need to find other ways to
do this. Long term the grab rectangle should be a widget but for now we
can just go about putting it beneath the text and set the selection type
if we did not get a color to use.