completion: use extra treeview for styling
We don't want this long term, but for now we can use it to get the theme
information out without too much trouble.

Long term, this will all be replaced with a new completion API that is
based on GListModel, so this is mostly just temporary.

Also, disable all the background styling for now since we can't do that
given current theming APIs. We want to use widgets for all of this.