stylescheme: simplify CSS provider usage
Now that we aren't using widget theming API to set the cursor state, we
can go ahead and just simplify the CSS provider to a single provider.

However, we also need to apply this to the gutter since the css provider
state is not inherited. That will change in the future, but until then we
can just apply it directly.

This fixes the background for the gutter.