Meson: Fix pkg-config file when a dependencies is a subprojects
Meson only expects strings and pkg-config dependency objects in the
`requires` argument, but when a dependency is built as subproject it's
an internal dependency object instead and that cause Meson to error.

It is rarely needed to put anything in 'requires', meson knows if public
libraries have a pkg-config name it can use. Also meson knows private
dependencies, no need to specify them.