jsx.lang: Highlight invalid characters as errors
This highlights } and > in text positions as errors. The JSX spec
already disallows this (along with { and <) but TypeScript 3.9 and Babel
are now enforcing this rule to be more comformant.

This also restructures the contexts so that /> outside of tags (i.e. in
text positions) will not end an element prematurely.