javascript.lang (and family): Highlight constructor in class bodies only
Previously, "constructor" would be highlighted as a built-in method in
all places where a property name is expected, e.g. in object literals
and object destructuring.

This change causes "constructor" to be highlighted inside class bodies
only.

Also do not highlight "constructor" if it is combined with other method
keywords (get, set, async, etc.).