completion: use gdk_window_move_to_rect()
A number of issues have been found when dealing with placement of the
completion window with multiple monitors as well as coordinate systems
outside of X11.

This switches GtkSourceCompletionInfo to use gdk_window_move_to_rect()
which has more robust implementations for platforms like Wayland.

Additionally, this requires a bump to GTK 3.24 for the necessary API as
it was removed from the private GDK vtable.

Signed-off-by: Christian Hergert <chergert@redhat.com>