assistants: add new assistants subsystem
This subsystem is meant to be a base for popup based tooling around a
GtkSourceView. It can be used for the underlying layer of completion,
interactive tooltips, error bubbles, snippet tooltips, and more.

A new completion engine will land first using this subsystem, but will
keep the GtkSourceAssistant type private similar to how GDK does for
internal (but exposed) types.