gtksourceregex: wrap GRegex through a shim
This creates an ImplRegex intermediate structure that we can use to switch
the implementation from GRegex into PCRE2. Doing so will not only be
faster, but also allow us to eventually migrate to using a JIT for the
regex implementation as well as avoid deprecations from GRegex in future
releases of GLib.

Other modules will eventually need to be ported to this, but focusing on
GtkSourceRegex would result in the largest gain from language specs.