contextengine: fix first-line-only in presense of BOM
This fixes an issue where we don't correctly highlight the first line with
a shebang if we found a BOM sequence. We need to check against an offset
of zero or one depending on the presense of a BOM (which is always a single
character GtkTextIter in the buffer).

I still see some issues with #168 in that the rest of the file is still
not highlighted until an incremental change comes in, but at least this
ensures the first line gets correctly highlighted.

Related #168