Bump glib dependency to 2.66
This is the version required by gtk4. This fix meson error on distro
like Ubuntu 20.04 LTS that has glib 2.64 but not gtk4, so the
configuration of gtk4 subproject fails because of too old glib version
found and it cannot fallback to glib subproject any more because
gtksourceview is using the system dependency already.