map: use new textview API to fix glyph alignment
This helps reduce the amount of grid lines we see across the font from
pixel alignment in text rendering. Our font is a bit special, in that we
know how it needs to be rendered to give the effect we're looking for.

Depends on https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/3690