map: mix foreground with background for general text
We don't want the map to detract from the document at hand, so it should
be a bit less vibrant for general purpose text. The specific tags we can't
do much about, and that seems to be fine for most code.

Instead of using alpha, we mix here so that we have faster rendering paths.