view: avoid alpha compositing for right-margin-position
We spend quite a bit of time drawing the right margin due to it being
an alpha blend above the background, text, etc. Not only does that alter
how the text itself looks, it also is very slow compared to alternatives.

Instead, we can pre-mix the alpha blended form with the text background
in cases where the text background is set. If we take care to draw the
margin on line backgrounds (such as current line), we have an output that
is very similar but much faster and more correct text rendering.