gutterrenderertext: cache common widths
Generally, this widget is used as the base for line drawing. Additionally,
it is generally used in a monospace fashion.

Until we have a reason to consider alternatives, we should aggressively
cache the widths and re-use them to avoid calculating layout sizes on
every line number.

This caches them by common sizes (1..5) so that we only do that once per
snapshot per width.