searchsearchcontext: use PCRE2 for regex searching
When porting to PCRE2, we only implemented the syntax highlighting engine
because there was a lot to port. That meant we used PCRE for search
contexts (via GRegex), but PCRE2 for highlighting.

Now that we have enough implementation in our PCRE2 wrapper, we can make
use of ImplRegex for search contexts too.

It's possible we have some fallout, but for what I've seen in the wild
thus far, the bases seem to be covered.

It's unlikely that we'll backport this to GtkSourceView 3.x or 4.x series.
However, if there is enough interest for someone to do that work and
really test it on platforms in the wild, we'd be willing to manage that
upstream with them.

Fixes #138