language-specs: use \N{U+} escape sequences
We seem to be having problems with \x{} on PCRE2 for reasons that are
still illusive. This works around the problem by using the \N{U+val}
style which appears to fix the issue.

The best I've come across is possible compilation settings in PCRE2
related to how these escape sequences work.

The previous commit now checks all language specs as part of the build,
so we can catch these sooner rather than later.

Fixes #202