view: remove border from current line highlight
This is problematic in that it only works when RGBA is used since it
essentially just draws the RGBA twice in some areas. We want to use
premixed colors as much as we can across the board and remove the use of
snapshot_background() too.

Therefore, we'll drop the extra border, and replicate it further in the
gutter so they match.

Doing so will not only give us faster rendering, but a fighting chance at
fixing #29 in GTK 4/GtkSourceView 5 based applications.