classic.xml: use fallbacks for line numbers
Now that we have fallbacks in place for this within GtkSourceView we
don't need to specify it in the style scheme. Doing so will improve the
situation where classic is used on dark or other non-standard themes.

Note that this is only for GTK4/GtkSourceView5 and forward, and will not
be backported to previous versions of GtkSourceView as the machinery is
completely different.

Fixes #29