map: improve slider position calculation and snapshotting
We get lots of warnings when using the overlay feature from GtkTextView
and it also requires doing more work than necessary. Instead, we can just
manage the child widget manually and update it on size requests to avoid
extra cycles on allocations (as well as errors about snapshotting without
an allocation).

Furthermore, we can simplify how we calculate the position.