vim: add GtkSourceVimIMContext
This adds a new GtkIMContext implementation that attempts to emulate
a modern Vim experience. It can be used with a GtkSourceView by
connecting it to a GtkEventControllerKey and adding it to a view.

This is an initial implementation and could use further work on
matching semantics with Vim. However it has a number of features
beyond what was implemented in GNOME Builder.

Various features supported by GtkSourceVimIMContext include:

 - Normal, Insert, Replace, Visual, and Visual Line modes
 - Support for an integrated command bar and current command preview
 - Search and replace
 - Motions and Text Objects
 - History replay
 - Jumplists within the current file
 - Registers including the system and primary clipboards
 - Creation and motion to marks
 - Some commonly used Vim commands

Fixes #233