gutterlines: improve yrange calculation
This could be slighly more expensive for the single line case, but we
really have to do that anyway or we won't take line-height into account.

Additionally, the first/last doesn't seem to be perfect in terms of
aligning with the pango run, but it does seem to be within a pixel.

Fixes #227