words: perform word completion from model
This moves the word completion work to a model rather than in the provider
itself as it results in much easier to understand code.

Additionally, it fixes an issue where we weren't seeing updates to the
results.

I'd like to note though, that there is still an initial scan timeout in
the buffer scanner to add words, which is 5 seconds. So if you're testing
this you'll need to wait for long enough for the word you just typed to
show up in the library or it will not be available for completion.

Fixes #181