stylescheme: add support for <metadata>
Language syntax files allow for <metadata> with <property> elements
within them. This can be handy to add additional non-critical information
to the style scheme which can be parsed by applications.

In particular, we'd like to use this to associate possible related
variants and other information in a somewhat flexible manner.

A new gtk_source_style_scheme_get_metadata() function has been added
similar to the equivalent in GtkSourceLanguage.

Related #220