vim: allow shortcuts to pass through from Normal
We want shortcuts to pass-through instead of being swallowed. However,
there isn't really a way to know for sure what will end up being a
shortcut and we can't test the GtkShortcutControllers up the hierarchy
because handle_event() is private API.

So instead just cheat and check for control/alt/super masks and empty
strings to determine if it's something relatively safe to pass along.