fileloader: use allocated chunk buffer
Previously, this had the chunk_buffer inline in the TaskData struct.
Unfortunately, since that creates an allocation that is offset from the
head of the struct, the value of having data inline is vastly overwhelmed
by all of the coordinating layers doing additional memory copies.

Buffers for data reads from storage should always be page aligned,
especially if you're reading multiples of the page size.