gutterrendererlines: remove pango markup usage
Now that the GtkSourceGutterRendererText can handle the changes for the
current line foreground color and weight, we no longer need to handle
this from the GtkSourceGutterRendererLines.

We can also save some time generating strings since we can always use the
helper now. Not to mention the parsing of markup too.