view: ignore current line during selections
If we have highlight-current-line set, but there is a selection, avoid
having both the line hightlight and the selection which is distracting.
The selection boundary itself is descriptive of positioning during an
interactive selection and it also avoids some double compositing with
weird color combinations.

If you really want the old behavior, we can discuss options to make that
possible, but I definitely don't think it should be the default here.