view: improve grid placement
This improves the placement of grid lines a bit by repeating at the whole
character size rather than the half-height. The previous form was more
likely to drift from the desired size just from odd-integer height of
a sized character.

Additionally, if we are not line-wrapping, we can use the lines yrange
to find a better half-grid-height so that we don't alias the grid boxes
across lines of text.