c.lang: place type keywords in separate style
This allows for styling the type keywords differently (but falling back to
the same behavior as before by referencing c:keyword) so that we can make
style schemes such as solarized more closely match their upstream counter-
parts.

For example, in Solarized (VIM at least), `if` and another keywords are
green except for typedef/struct/enum/union which are yellow.