json.lang: Fix styles for floating point values
This fixes a bug where characters from the start of a floating point
value up to the second character before the decimal point or exponent
would be highlighted as integer values. This was noticeable in
themes that differentiate between the two numeric types.

The json test file has been updated to verify this change.

Fixes #250.