Adwaita.xml: more updates to Adwaita style
This still has the full palette I'm working with until we've modernized
the Adwaita-dark scheme to match.

This also raises the default text contrast.

Some palette colors have been extended so that we can keep things that
fit with the (very limited GNOME HIG) colors but are dark enough to be
used (or mixed with other colors in the palette) so that we have more
options. For example we've added violet and teal which are not part of
the original palette, but are mixed of other palette colors.

Others have additional dark colors which were taken by finding HSL values
for various colors and tweaking them to add darkness.