view: reduce chances for textview overdraw
Currently there is the chance that a textview could cause the GPU to render
large color regions at least 4 times.

 1) The window background
 2) The .view background for GtkSourceView
 3) The view-style applied to "textview text" background in many themes
    which inherited from the old Adwaita/Default CSS (now fixed in both).
 4) The right margin area

If we force disable #3 we get the same results but it also allows us to be
able to snapshot some render nodes from gtk_source_view_snapshot() before
any GSK transforms are applied. This will be important because it improves
the chances that the GSK render node diff code will elide the large
rectangular regions.