view: allow right-margin snapshots to be elided
We can elide much of the right-margin color rectangles if we snapshot
it from gtk_source_view_snapshot() rather than
gtk_source_view_snapshot_layer() where the GSK transform has already
applied a translation.

We could not do this before the previous commit because it was possible
that "textview text" had erroneously styled the background to a color
from both the Default and Adwaita themes (now fixed upstream). However,
now that we force override that, we can draw here and have it placed
correctly and get the same rendering, but with much less pixels damaged
in the GPU w/ scissor rect applied.