tests: start on revamped test-widget
The goal here is to be more similar to how applications, and particularly
editors, are going to be rendering content (large editor space, taking
most of the window child).

Additionally, we want to be using CSD as that is more prevalent and we
need to be sure our rendering performance is good with windows which can
be, at least partially, transparent.