snippet: add gtk_source_snippet_new_parsed()
This is a function that can create a new snippet by parsing the snippet
format used in bundles. It happens to be able to parse a decent amount of
language-server-protocol snippets as well, which should be what we aim
for long-term to maximize re-use in IDE tooling such as Builder.