language-specs: add suggested suffix/name properties
For a number of languages, we can help text editors choose a reasonable
suffix for the filename using metadata properties. GNOME Text Editor, for
example, will use this to prime the save changes dialog.

Use gtk_source_language_get_property(lang, "suggested-name") first to get
the suggested name, and if that is NULL, use "suggested-suffix" with a
generated prefix (or leave it empty with the cursor set before the suffix
in the filename entry).