build: Also try finding libxml2 with CMake
For Windows builds of libxml2, one may be using its supplied NMake Makefiles
to build libxml2, which do not have complete pkg-config files, or use CMake,
which optionally crafts its pkg-config files that may need some tinkering for
proper use.

Instead, save from such issues to try finding again libxml2 with CMake, which
actually manually search for libxml2's headers and libraries like what we used
to do, for us, if a proper pkg-config file cannot be found by Meson, before we
resort to the fallback.

If at one point we require Meson 0.60.0 (or so) or later, we can just put the
package strings that we want to search for in a form like
['libxml-2.0', 'LibXml2'], but let's just do this for the time being.

The "LibXml2" format follows the CMake docs at [1].

[1]: https://cmake.org/cmake/help/latest/module/FindLibXml2.html