build: Look for Vulkan
...and include it in core_deps if it was not found via pkg-config.  The reason
for this update is because we have no way to find out from the Gtk4 dependency
whether Vulkan is enabled for Gtk4, and so neglect to include the header dir
from the Vulkan SDK, which will break the build, unless Vulkan was found via
pkg-config when building Gtk4.

If Vulkan was found in ways other than pkg-config, include it as a dependency
of core_deps so that the build can find its headers.