testsuite: backport test-language-specs
This adds a test to the testsuite that tries to load every language-spec
that is bundled with GtkSourceView. It generally catches errors that would
cause GtkSourceView to throw warnings or critical errors when loading.

Currently, it fails due to latex.lang missing a def:text style which we
can cherry-pick the fix from 5.x for.