sourceview: improve input-method detection for indenters
This attempts to improve the situation where we need to detect if an input
method filtered our keypress but really inserted what we expected it to.
This can happen with ibus over wayland stealing input like > or / and
marking it filtered (but going through "commit-text").

To make this work, we have to introspect a bit more by trying to convert
our keyval to a unicode character, and then checking equality with what
was just inserted at the previous character.

Fixes #282