buffer: rearm timers instead of remove/add
For things that can happen at high-velocity, it is cheaper to change the
ready time on a GTimeoutSource than it is to remove/add a new GSource.

While loading should probably disable bracket highlights altogether, this
is a reasonable thing to have anyway for code paths that don't use helpers
to load/unload buffer contents.

Related GNOME/gnome-text-editor#444