bufferoutputstream: disbale cursor-moved while writing
We can disable the cursor-moved completely while writing to the buffer,
particularly to improve the situation when fallbacks are used. This helps
reduce how much callback work is done by applications.

When the block count reaches zero (which could happen in the epilog of this
function), the buffer will emit cursor-moved so that applications can
handle this as if it were a single change.