hover: keep GSource around for re-use
One minor value of this is that we can set the name without incurring the
GMainContext/Source locks multiple times. Just set it (without those locks)
after creation so long as we have a GSource pointer.

Additionally, and more importantly, if we keep the GSource around, we can
change the ready time on the GTimeoutSource instead of remove/add flows
which are not ideal when you do as many updates as we do.