snippetbundle: give search listings access to tooltips
We could probably limit these if we wanted, but it's unlikely to be
worth it compared to just memcpy() as the complexity would be greater
than the memcpy costs when you have to look at every element anyway.

A better long term solution would be to have a tooltips structure that
we can reference (other than just the array) which we don't want to
possibly mutate.