completionlist: avoid grabs in completion list
This follows suit of other GtkSourceAssistant by removing the grabs that
would otherwise be taken by the completion list and instead uses
GtkPopover:autohide=FAlSE.

This allows us a bit more flexibility in how events are routed but also
ensures that input doesn't continually enter/leave the source view on
every display of the list.

It also simplifies the grab management for the native surfaces since they
no longer need to track crossings and break grabs.