assistant: let caller know if position changed
The hope for this originally is that we could use the result of this
function to determine what level of widget invalidation we needed on
the GtkSourceView to ensure we had an allocation for the next snapshot.

In practice, it didn't matter because either way we need to do a full
gtk_widget_queue_resize(). However, it does attempt to avoid invalidating
some state in the popover so it's useful enough to keep it around so we
know explicitely when we've changed things.