completionlist: use 1 for target position width
Using 0 was not really useful because it meant we could never compare the
rectangle with a previous value. gtk_popover_get_pointing_to() will always
return 1 for a width if it was set using a zero-width rect.

Using 1 means we have a stable value that can be compared across a setter
and getter for the GtkPopover.