assistant: gtk_widget_queue_resize() when updating position
When we update our position from the next frame cycle, we need to also
ensure that we measure/size_allocate for the updated popover. Otherwise we
risk having warnings about allocations not being up to date.

This fixes a bunch of flickering in the source view while using completion
popovers which rapidly update.