solarized: use error underline for def:error
Using a bold red color for errors might have matched the original styles
of Solarized when it was captive to a terminal, but it does not very much
match modern IDEs and editors when it comes to what the pango renderer
can do.

Instead, use the expected error styling but color to match the Solarized
palette.