completion: fix activation with empty bounds from trigger
This can easily happen if you have an operator that caused the completion
like "." as the symbol chars will break at the "." and leave us with
bounds that are identical.

If we track that the interactive completion came from a trigger, we can
ignore this and assume we want to begin completion immediately.