hover: disable context menus from hover assistant
This is a workaround until we can be sure that we can show popovers from
the hover assistants. Currently, that can break input pretty badly, so
this is in place to help avoid that issue by users.

This basically makes it so GtkLabel (and potentially others) cannot get
their sequence event to show the context menu since we don't have other
mechanisms to prevent it.