vim: ignore gtk_text_view_im_context_filter_keypress
There are cases while in INSERT mode that we need to handle to switch
to REPLACE as well as switching to literal input.

These all currently require the Control key or Escape along with the
Insert key. Filtering them would return FALSE, but unforutnatley if
they are not those, calling gtk_text_view_im_context_filter_keypress()
will cause the text to be committed before further GtkEventControllerKey
instances could process the input.

That breaks other GtkSourceView features such as indenters that might
trigger on the key-press. For example, Builder can reformat function
arguments on `)`. Without this, that code and other key-press handlers
would be completely skipped while in INSERT mode.