snippets: fix :enable-snippets without completion provider
If we are expanding snippets, and there is no completion provider used to
show the available snippets, commit 04683f367 broke the ability to reach
the code which would try to expand.

This attempts to preserve the existing behavior to make that work while
also doing the intention of commit 04683f367 to hide the informative when
it is no longer necessary (and only if a snippet was touched).

Fixes #302