init: call gtk_source_init() from library static constructor
This uses static constructors like we do in a few other GNOME libraries to
initialize state up front before main() is called, or when the library is
dynamically loaded via dlopen().

This allows language bindings to omit the call to GtkSource.init().

We couldn't do this in GtkSourceView 4.x and previous due to gtk_init()
requiring argc/argv parameters which is no longer the case. Additionally,
repeated calls to gtk_init() are ignored.