assistant: update popover positioning immediately
Now that a lot of other layers appear to be fixed, we can ramp back up the
number of repositions we do so they don't fallback to the next frame
cycle.

This improves the "jitter" that you feel when updating the completion
results in Builder.

You may want to have an up to date GTK if you're seeing issues with this
as lots of invalidations are being fixed in lower layers.