vim: avoid small deletions in filter command
Removing text and then inserting it turns out to be really expensive when
there are lots of signal handlers. Instead do what we do elsewhere and
remove all of the text only to re-insert it/indent the line one by one.

This fixes some issues in Builder where reformatting with = on large
buffers can bring things to a hault.