vim: Visual replace text
This commit makes Vim emulation replicate the behavior of "putting" text
while in visual mode. When the user presses "p" with a selected range
the range's content will now be replaced by the contents of the current
register.

Furthermore, this commit also adds the same feature for the system
clipboard by using "Ctrl + Shift + v", similar to what is present in
insert mode.